Fix for issue #130
There may be a more elegant way to do this, but the gist is that we pluck all of the dates for the event's shifts, concatenate them together, and then use array_unique to remove duplicates. Then, instead of testing for a start_date, we just see if the current $date is present in that list.